+++ src/main/kotlin/com/github/zijing66/dependencymanager/services/GradleConfigService.kt
@@ -106,7 +106,7 @@ class GradleConfigService(project: Project) : AbstractConfigService(project) {
         configOptions: ConfigOptions,
         path: String,
         pkgData: PkgData,
-        onDirFound: (File, String, PkgData) -> Unit
+        onResultFound: (File, String, PkgData) -> Unit
     ) {
         val targetGroupArtifact = configOptions.targetPackage.takeIf { it.isNotEmpty() }?.split(":")
 
@@ -150,7 +150,7 @@ class GradleConfigService(project: Project) : AbstractConfigService(project) {
         }
 
         if (shouldInclude) {
-            onDirFound(pkgData.packageDir, matchType, pkgData)
+            onResultFound(pkgData.packageDir, matchType, pkgData)
         }
     }